fix parenthesis in arithmetic functions fix duckdb table order_details

+++ northwind/src/main/resources/ai/koryki/databases/duckdb/northwind/db/tables.sql
@@ -106,9 +106,9 @@ CREATE TABLE employee_territories (
 CREATE TABLE order_details (
     order_id smallint NOT NULL,
     product_id smallint NOT NULL,
-    unit_price NUMERIC(18,6) NOT NULL,
-    quantity NUMERIC(18,6) NOT NULL,
-    discount DECIMAL(5,4)  NOT NULL
+    unit_price DOUBLE NOT NULL,
+    quantity SMALLINT NOT NULL,
+    discount DOUBLE NOT NULL
 );
 
 --
